Move misc kotlin2cpg logic into own files
some files have gotten a bit too hard to read
