+++ joern-cli/frontends/kotlin2cpg/src/main/scala/io/joern/kotlin2cpg/passes/ConfigPass.scala
@@ -1,6 +1,6 @@
 package io.joern.kotlin2cpg.passes
 
-import io.joern.kotlin2cpg.{FileContentAtPath, Kt2Cpg, KtFileWithMeta}
+import io.joern.kotlin2cpg.{FileContentAtPath}
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.NewConfigFile
 import io.shiftleft.passes.{DiffGraph, IntervalKeyPool, ParallelCpgPass}
