+++ joern-cli/frontends/kotlin2cpg/src/test/scala/io/joern/kotlin2cpg/SourceFilesPickerTests.scala
@@ -1,5 +1,7 @@
 package io.joern.kotlin2cpg
 
+import io.joern.kotlin2cpg.files.SourceFilesPicker
+
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.BeforeAndAfterAll
